+++ R/esp_get_capimun.R
@@ -33,6 +33,9 @@
 #' all the municipalities of that level would be added.
 #' @export
 #'
+#' @note While \code{moveCAN} is useful for visualization, it would alter the
+#' actual geographical position of the Canary Islands.
+#'
 #' @examples
 #' \donttest{
 #' # This code compares centroid of municipalities against esp_get_capimun